Add Contracts model to Prisma schema
A new Prisma model called Contracts has been added to the Schema. This model includes various properties related to contracts, such as Serial, NameEntity, Department, and more. The addition of this model will allow for a more structured and organized approach to handling and manipulating contract related data in the application.