Create 'Contracts' table and XServer module for SQLite database
Created a new module 'XServer' for interacting with a SQLite database. Also added a new 'Contracts' table to the database with multiple fields. These changes also included setting up the necessary dependencies in package.json, adding a README for instructions on running the module, and setting up a tsconfig.json for TypeScript configuration.