Add EntityContract component to app component
Added a new EntityContract component for displaying contract entity information. This new component, which includes the name, nit and central entity of the contract, provides more structure and unifies the way this information is presented. This change simplifies the app.component.html by replacing repetitive html code with the new EntityContractComponent.