Add ListContractsComponent to show contracts
Created a new component named ListContractsComponent that fetches and displays all the contracts in a formatted list. The list markup from AppComponent was moved to this new component in order to keep AppComponent minimal and generic. The ListContractsComponent component is now loaded on 'contracts/all' route, this approach enhances code modularity and readability.